ref(conversations): Drop New suffix from redesigned components
Now that the flag-off components are gone, rename the redesigned
components and files to their base names (e.g. ConversationsTableNew ->
ConversationsTable) and fold the thin conversationDetail wrapper into the
page it delegated to.

Pure rename; no behavior change.

Refs TET-2742
