+++ static/app/views/explore/conversations/overview.tsx
@@ -23,7 +23,7 @@ import {
   ExploreBodySearch,
 } from 'sentry/views/explore/components/styles';
 import {TraceItemSearchQueryBuilder} from 'sentry/views/explore/components/traceItemSearchQueryBuilder';
-import {ConversationsTableNew} from 'sentry/views/explore/conversations/components/conversationsTableNew';
+import {ConversationsTable} from 'sentry/views/explore/conversations/components/conversationsTable';
 import {SaveConversationQueryButton} from 'sentry/views/explore/conversations/components/saveConversationQueryButton';
 import {useShowConversationOnboarding} from 'sentry/views/explore/conversations/hooks/useShowConversationOnboarding';
 import {ConversationOnboarding} from 'sentry/views/explore/conversations/onboarding';
@@ -134,7 +134,7 @@ function ConversationsOverviewPage() {
           ) : showOnboarding ? (
             <ConversationOnboarding onDismiss={refetchOnboarding} />
           ) : (
-            <ConversationsTableNew />
+            <ConversationsTable />
           )}
         </Stack>
       </ExploreBodyContent>
